# Conflicts:
#	build.gradle

+++ jsettlers.logic/src/test/java/jsettlers/integration/replay/ReplayValidationIT.java
@@ -59,15 +59,14 @@ public class ReplayValidationIT {
 
 	@Test
 	public void testIfReplayIsEqualToOriginalPlay() throws IOException, MapLoadException, ClassNotFoundException {
-		final float targetTimeMinutes = 60f;
+		final int targetTimeMinutes = 60;
 		MapLoader map = MapUtils.getMountainlake();
 
 		ReplayUtils.PlayMapResult directSavegameReplay = ReplayUtils.playMapToTargetTimes(map, targetTimeMinutes);
 		assertDirectSavegameReplay(1, directSavegameReplay);
 		MapLoader savegame = directSavegameReplay.getSavegames()[0];
 
-		MapLoader replayedSavegame = ReplayUtils.replayAndCreateSavegame(directSavegameReplay, targetTimeMinutes,
-				REMAINING_REPLAY_FILENAME);
+		MapLoader replayedSavegame = ReplayUtils.replayAndCreateSavegame(directSavegameReplay, targetTimeMinutes, REMAINING_REPLAY_FILENAME);
 
 		// compare direct savegame with replayed savegame.
 		MapUtils.compareMapFiles(savegame, replayedSavegame);
@@ -75,7 +74,7 @@ public class ReplayValidationIT {
 
 	@Test
 	public void testIfSavegameOfSavegameEqualsSavegame() throws IOException, MapLoadException, ClassNotFoundException, InterruptedException {
-		final float targetTimeMinutes = 30f;
+		final int targetTimeMinutes = 30;
 		MapLoader map = MapUtils.getMountainlake();
 
 		ReplayUtils.PlayMapResult directSavegameReplay = ReplayUtils.playMapToTargetTimes(map, targetTimeMinutes);
@@ -104,7 +103,7 @@ public class ReplayValidationIT {
 	@Ignore
 	@Test
 	public void testReplayForSavegame() throws IOException, MapLoadException, ClassNotFoundException {
-		final float[] targetTimeMinutes = new float[] { 30f, 60f };
+		final int[] targetTimeMinutes = new int[] { 30, 60 };
 		MapLoader map = MapUtils.getMountainlake();
 
 		ReplayUtils.PlayMapResult directSavegameReplay = ReplayUtils.playMapToTargetTimes(map, targetTimeMinutes);
@@ -112,7 +111,7 @@ public class ReplayValidationIT {
 
 		MapLoader[] savegames = directSavegameReplay.getSavegames();
 		for (int i = 0; i < targetTimeMinutes.length; i++) {
-			float targetTime = targetTimeMinutes[i];
+			int targetTime = targetTimeMinutes[i];
 			MapLoader savegame = savegames[i];
 
 			MapLoader replayedSavegame = ReplayUtils.replayAndCreateSavegame(directSavegameReplay, targetTime,
